+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added `get_int_kinetic`, `get_int_nuc_attr`, `get_int_hmo` getters for one electron integrals (#243)
 - Added `TwoElectronIntegralElement`, `MOTwoElectronIntegral` and `TwoElectronIntegrals` for structuring two electron integral information (#243)
 - Add the keyword `dumpactints` to `BlockOutput` (#245).
+- Added fallback keyword argument for `get_final_energy`, `get_gradient`, and `get_structure` that allows to parse these properties from the `.out` file if no JSON output is available (#237).
 
 ### Changed
 - Refactored methods from Runner into BaseRunner (#193)